Fixed bug in the HcRhPortStatus bit set
The issue with the HcRhPortStatus not setting the PortEnableStatus(PES) when
the PortResetStatusChange(PRSC) bet was cleared by setting it was because of
the usage of set_bits and clear_bits. The set_bits preserves the bits
and writing it back would cause the reset of other parameters wherever
the bit was set. To fix this the helper macros just writes only that bit to be
set in the register without any preservation of the previous which would
cause unintented clearing of bits. The above works because writing a '0'
has not effect.