Added unsigned suffix to GPIO_BA, GPIO port blocks, GPCON, GPDAT, GPUP and MISCCR_OFF.
Added USB_EN_GPIO bit position for the USB device.

Setup the GPC bit 5 control to output.

Disabled GPC5 pull and set the output to 0.

 * The above fixes the problem of the CSC bit set for Port2. After
   disabling the USB_EN which disables the USB device the CSC Port2 gets
   cleared.

Modified the reset of the HC_COMMAND_STATUS_REG by just writing the HCR
bit rather than read modify and write. Modified the same for CLF.

The above is done because the HCD can set and the HC can clear. Writing
a clear by the HCD is ignored by the HC.