Added memset function to set the buffer with the set byte value.
Created a generic memcpy and memset.

 * Made the dest,src of memcpy and buff of memset void pointers
 * Typecast to uint8_t * from void * inside the function logic

Chained TDs to each other in init_td with the last TD set to 0.

Added a get_free_tds to get free TDs.

Added reclaim_free_tds to reclaim the TDs which are being unused and
free.

Removed set_usb_desc_req_buffer function to fully utilize the macro of
the same functionality.

Added comment max_packets for the extra 1 packet.

Added additional comments on HeadP and TailP settings of the ED.