Renamed get_free_tds to alloc_td.
Renamed reclaim_free_tds to reclaim_unused_td.

Initialized td_info->free_td_size in init_td.

 * Due to not initializing the alloc_td was returning 0. This was
   because free_td_size was 0 and hence no free TDs.

Added documentation on the different steps of alloc_td and
reclaim_unused_td.

Added termination of the TD list with the next_td pointing to 0 in
alloc_td

Introduced temp_td in alloc_td which is used for traversal as well as
return of the free_td_head.

Introduced len which keeps a count on the amount of TDs reclaimed in
reclaim_unused_td.

 * Added the len to the free_td_size to keep track of the amount of free
   TDs left.

Updated set_setup_descriptor to use the allocated TDs.

 * Allocated with hardcoded 5 TDs for testing.
 * Updated the SETUP Request TD using the allocated TD.
 * Updated DATA packet TDs using the allocated TDs.