Reverted send_td_request_pkt to minimal settings.
send_td_request_pkt has been reverted to minimal settings. This is due
to the fact that the status of the request has to be checked after
sending the request TD and data TD if any. Only after this the status
TD has to be sent. The previous code was written with the assumption
that the TD status has to be checked immediately after the request
header is sent and not after sending the data TD.